Add `XkbCommonCompose` struct for compose module
This commit adds a new struct that only includes libxkbcommon functions
from the compose and dead-keys support module and removes these
functions from `XkbCommon`. It is now possible to use an
`XKBCOMMON_HANDLE` with libxkbcommon 0.4.3 which does not include this
module. Availability of this module can be checked with
`XKBCOMMON_COMPOSE_OPTION`.

Resolves #1