chore: load sonames first
The full soname, with the version number of the so file should be tried before the one without a version number.

If this is not done, the latest version, which probably is no longer compatible, might be used instead of the version that the library was compiled against and cause problems.

The previous behaviour was also problematic when trying to create a classic mode snap for Neovide. libxkbcommon.so is not installed unless you install the dev package. So then it's not found in the snap itself, but might be found on the host and loaded from there, which can cause problems. When using the full version libxkbcommon.so.0 it's found inside the snap and loaded from there as it should.